+++ src/Producers/ProducerBuilder.php
@@ -102,7 +102,7 @@ class ProducerBuilder implements CanProduceMessages
      * @param array $headers
      * @return \Junges\Kafka\Producers\ProducerBuilder
      */
-    public function withHeaders(array $headers): self
+    public function withHeaders(array $headers = []): self
     {
         $this->message->withHeaders($headers);