+++ src/Support/Testing/Fakes/ProducerBuilderFake.php
@@ -122,7 +122,7 @@ class ProducerBuilderFake implements CanProduceMessages
      * @param array $headers
      * @return \Junges\Kafka\Support\Testing\Fakes\ProducerBuilderFake
      */
-    public function withHeaders(array $headers): self
+    public function withHeaders(array $headers = []): self
     {
         $this->message->withHeaders($headers);